Scenes: Record previously visited scenes
Before we can implement preconditions for scenes, i.e. requiring that
the player has visited scenes X and Y before being allowed access to
scene Z, we need a mechanism for tracking the scenes which the player
visits.  This patch implements the basic tools we can use for this
purpose.

We begin by introducing a ‘VisitedScenes’ table which maps the names of
scenes to booleans indicating whether the player has gone through them.
In order to populate this table we modify the ‘set-scene’ instruction,
the lowest-level point in the engine where we switch scenes, to update
data in that ‘VisitedScenes’ table.

The result is that changing between scenes builds us map that allows us
to quickly determine of the player/reader has seen a given scene or
not.  Technically speaking, it only tells us if the player has entered
that scene, and it is not in indications of how much content the player
witnessed within the scene.  Nonetheless it provides a strong
technological starting part for implementing such future features and
preconditions for being allowed to view scenes, because we can use these
changes to determine if the player has visited the prerequisite scenes
necessary to unlock access to any given scene.