Scenes: Remove the ‘name’ property
The original intent of Scene objects having a name property was to
help provide better debugging output.  However, we are moving more and
more in the direction of the variable names for Scene objects as
representing the name for that scene.  For example, if we have a
dialog script with the following

    Meeting = Scene {
        …
    }

then ‘Meeting’ is the name of that scene.  At a technical level
‘Meeting’ is the key in the LNVL.ScriptEnvironment table which points
to the LNVL.Scene object representing that scene.  This makes the
variable names the de facto “names” of scenes.  And so for that reason
we remove the long-standing ‘Scene.name’ property, an optional
property that we could define when creating a scene.  We only ever
used that property to help create debugging output.  But we can
achieve the same quality of diagnostic output without needed that
optional object property.