Keep track of the history of visited scenes in order
Some games will want to know exactly how a player traveled from scene A
to scene B.  This patch implements a table that provides exactly that
level of tracking.  The table is accessible outside of LNVL so that
games implementing the engine can make use if it as they see fit.