Remove opcode processors from the public API
We always invoke the processor function for an opcode by calling the
process() method on an Opcode object.  That method uses the name of
the opcode to find the processor in a look-up table.  This patch makes
that table private because the process() method is all we need, not
access to the processor functions themselves.