Characters: Accept color names for text and border colors
When creating a character in a dialog script the Character constructor
will treat strings for properties like ‘textColor’ as hex color
strings, e.g. ’#33cfaa’.  This fails to account for situations where
the string explicitly names a color, e.g

    Eric = Character { textColor = "Black", ... }

This patch fixes this bug by first testing if a given color string is
a valid key in the LNVL.Color table.  If so then we use the table of
RGBA values mapped to that name.  Failing that we revert to the
original behavior of treating the string as a hex color value.