Add a debugging function to dump the script environment
This gives us a way to see all of the variables in the
LNVL.ScriptEnvironment table along with their types.  This does not
give us specific type output when the values are tables, e.g. we do
not see whether a table represents a Scene object versus a Character.
However, this basic amount of output lets us quickly see what
variables exist in the environment, which is useful for debugging.