main: Dump the script environment after loading scripts
This modifies love.load() so that we dump the script environment if we
are running in debugging mode.  The intent is to help us track down
current bugs in example scripts that change scenes, as they are
crashing with errors about not being able to find scenes in the script
environment.