Fix the indentation for the file-header comment in LNVL.lua
One line was accidentally indented in by two spaces, which is a
mistake that I have seen happen more than once via Emacs.  It only
seems to happen with block comments that have ‘--’ before each line,
so this patch simply removes those.

This patch only changes the formatting of a comment.  It does not
change any actual code.