[Feature] Allow injecting data into LNVL dialog scripts
This patch provides a large new feature to LNVL.  It is desirable to
provide programmers with a means for injecting data into LNVL scripts
from outside of the LNVL sandbox.  In fact, it is mandatory if we want
LNVL to be viable for use within larger games.

To accomplish this we first introduce a new Context class.  A Context
object is a wrapper around a lookup-table that we use to augement the
environment when loading an LNVL dialog script.  We can use these
objects to define variables which become accessible in those scripts.

To create these variables we modify LNVL.LoadScript() so that it
accepts any number of optional Context objects and modifies the script
environment before performing any loading.  We merge each context into
the global environment for scripts, i.e. LNVL.ScriptEnvironment, so
that all future dialog scripts have access to the information of each
context.  If there are name conflicts between contexts then the most
recently loaded context takes precedence.  There is no way to tell
from the API when this happens.

Finally the patch provides a new test case demonstrating the
functionality, which required slight change to the demo ‘main.lua’
file since we need to create a Context there before loading any
example script.