[Feature] Prevent overwritting of core LNVL functions
It is possible for a player to accidental name a character with the
name ‘Character’, or anything similar.  This can potentially blow up
the entire engine by robbing the user of fundamental functions
necessary to write scripts.  This patch introduces a table of reserved
keywords in which we store the names of such things as object
constructors, thereby preventing script authors from blasting over
important LNVL functionality.