[Feature] Allow setting script variables in dialogue scripts
This patch lays the foundation for being able to change the value of
variables in the script environment during the processing of a
dialogue script.  However, it is currently flawed as the changes to
LNVL.ScriptEnvironment do not propogate as expected.  The patch
introduces a new opcode and instruction pair: ‘export-variable’ and
‘set-data’, respectively.  The new function Scene.export() is
responsible for creating that opcode and this the instruction.  The
example script in this patch demonstrates the syntax for exporting
variable values.