[Feature] Implement ’Get’ for dialogue scripts to load variables
This patch provides a new ’Get’ function for use in dialogue scripts
which will load a variable from the script environment.  For example,
if a script contains

    Get "secretMessage"

then the corresponding instruction will render the text assigned to
LNVL.ScriptEnvironment["secretMessage"].  This helps complete the
full-circle between being able to both import and export values
between LNVL and the game hosting it.

The downside to this approach, however, is that it does not allow
things such as string concatenation, which may be something we want in
the future.