[Feature] String interpolation in dialogue scripts
It is desirable to insert the value of variables into strings of
dialogue within scripts.  For example, we may wish to write a line of
speech where a character says something defined in a variable, a
variable with a value that may change depending on external factors
(e.g. something in the game wrapped around LNVL).  This patch
implements that functionality.

References to variables look like this:

    START = Scene {
        "Variable FooBar has the value <<FooBar>>",
    }

The syntax for interpolation is <<...>>.  LNVL will look for the
variable named within the braces and insert its value.  That variable
must exist in the script environment, i.e. LNVL.ScriptEnvironment.  It
is possible to define variables for interpolation within dialogue
scripts themselves or to provide them by using a Context object.

See the file 'examples/20-StringInterpolation.lua' for more details
and how interpolation works.