[Fix] Partially enforce the return value requirement for opcodes
The documentation says that opcode processors must return a table:
either an 'LNVL.Opcode' object (a table), or an array of said objects
(also a table).  However, no code checks the return value of
processors for this requirement.  This patch modifies the
Opcode:process() method to ensure that processors return tables;
although it does not further check metatables.  But this enforces some
of the type-value requirements for those processors.