[Feature] Support functions as preconditions for scenes
Currently the preconditions for a scene can only be a list of
pre-requisite scene names, as strings.  With this patch developers can
use functions in the preconditions table of a scene.  A function must
return a boolean.  If true that means the precondition represented by
that function is satisfied.  If false, then the precondition is
unsatisfied and LNVL will produce an error message.

The code for processing preconditions is not part of a local utility
function that is called from within the 'set-scene' instruction.

This patch updates the 'examples/17-ScenePreconditions.lua' script to
show how to use functions for preconditions.  Futhermore, it improves
the documentation for the 'Scene.preconditions' property to reflect
support for functions.