[Feature] Allow ‘ChangeTo’ to accept functions for creating branches
Right now the ‘ChangeTo’ command in scenes only performs an
unconditional jump to another scene.  That design makes it impossible
for branching out into different scenes.  This patch addresses that
limitation by teaching ‘ChangeTo’ to also accept a function which
returns the name of a scene, to which the story jumps.  Unlike using
plain strings, however, the function can contain arbitrary logic and
return the names for different scenes based on said logic.  There
rewrite of

    examples/08-ChangingScenes.lua

demonstrates this new functionality.  This allows script authors to
jump to different scenes and story branches based on criteria of their
choosing.