[Feature] Make Scene.changeTo() the public API for switching scenes
Outside of dialogue scripts the only way to change scenes, i.e. affect
the value of LNVL.CurrentScene and related variables, is to use
LNVL.LoadScript().  But that function only changes scenes when loading
a script that defines a ‘START’ scene.  There is no public API
function for changing to an arbitrary scene.

This patch provides that missing functionality by re-coding
Scene.changeTo().  That function now accepts the name of a scene and
can be used to arbitrarily change the current scene.  Previously the
function acted as the implementation for the ChangeToScene() function
within dialogue scripts.  This patch moves the old logic of
Scene.changeTo() into the definition of ChangeToScene(), preserving
its existing behavior.  Now the logic of Scene.changeTo() contains
most of the error-checking and bookkeeping we did inside of the
‘set-scene’ instruction.  This reduces some copy-paste code and lets
us put all of the logic for changing scenes (outside of dialogue
scripts themselves) into one place.

Finally, this patch updates the code in LNVL.LoadScript() to make use
of the new Scene.changeTo(), and the test file

    examples/17-ScenePreconditions.lua

is slightly modified to ensure that the new code correctly marks
‘START’ as a visited scene and adds it to the scene history.