[Test] Add a test and documentation for temporary characters
Sometimes we want a character only for a single scene, e.g. an
unimportant NPC.  Normally creating a character adds an entry into the
LNVL.ScriptEnvironment table.  However, as this patch demonstrates,
simply assigning that character a value of 'nil' at the end of a
script is enough to make them temporary, restricting them to the
dialogue script in which we create them.