[Feature] Run functions as scene content in the script environment
We allow users to write functions as parts of scenes, functions which
can return a string or opcodes representing content.  The original
design assumes that those functions should have access only to the
script environment but the implementation does not restrict those
functions that way.  This patch explicitly uses setfenv() for any
scene-content functions to execute them in LNVL's script environment,
and updates the documentation in the Howto accordingly.