[Docs] Clarify the purpose of LNVL.ScriptEnvironment.__newindex()
The comment above the metatable definition does not precisely state
why we define a __newindex() metamethod for the script environment.
We do so in order to prevent code from redefining any value in the
environment whose key is part of the ReservedWords table.  This patch
expands on the comment by explicitly describing the relationship
between the metamethod and the ReservedWords table.