[Fix] Prevent crashing when a dialog script has an empty Scene
Let's say we have a dialog script and this is the only content:

    START = {}

It defines the mandatory 'START' scene but contains no content.
Because no content results in no opcodes LNVL will crash on various
function calls we make within Scene.DefaultHandler().  This patch
addresses the issue by making the handler return immediately if there
are no opcodes, making the entire process a no-op.  Furthermore, the
constructor now produces a warning if we create a Scene object with no
opcodes.

We still allow empty Scenes, though, because for the time being we use
them in some testing.  But later it is likely that we will make it a
fatal error for Scenes to have no opcodes.