[Feature] Add a unique ID to all Scene and Characters for debugging
Given two Scene or Character objects we can compare them directly to
see if they are the same, i.e. do they point to the same place in
memory, but we cannot distinguish between them in a human-readable way
that is conducive to debugging.  For example, if we compare two Scenes
we expect to be the same and it turns out they are different, we have
nothing to readily display allowing the developer to tell which Scene
is which in the code or where it comes from.  This patch addresses
this issue by giving all Scene and Character objects IDs.  The ID
comes from the dialog script defining the Scene or Character; when we
load such scripts into LNVL we store all of the Scenes and Characters
as values in the `LNVL.ScriptEnvironment` table, where the keys of
that table are strings naming the variables which represent those
objects in the *original* dialog script defining them.  This patch
uses the keys in the script environment table as the IDs, meaning the
variable name from the original script becomes the ID.  For example,
let's say we have a script with the following:

    Eric = Character { ... }

    START = Scene { ... }

We this would result in the following objects being in the script
environment with the given `id` properties:

    LNVL.ScriptEnvironment["Eric"].id  == "Eric"
    LNVL.ScriptEnvironment["START"].id == "START"

These IDs are meant only for debugging and any code outside of LNVL
itself should not rely on them.