[Feature] Allow a custom handler for changing scenes
Developers using LNVL may wish to customize what happens when the
engine changes scenes.  For example, they may wish to use a custom
graphical transition effect on the screen.  This patch restructures
the `LNVL.ChangeToScene` function so that it can be defined via a
custom handler: `LNVL.Settings.Handlers.ChangeScene`.  The original
implementation of `LNVL.ChangeToScene` is now the default handler in
the event there is no custom handler available.  This has no immediate
changes on the behavior of LNVL, but simply opens the way for more
refined control when integrating the engine into a game.