[Feature] Protect against dialog scripts redefining reserved keywords
We use the `ReservedKeywords` table to keep track of things which
dialog scripts should not be able to redefine, e.g. constructor
functions for characters and scenes.  However, the existing mechanisms
for preventing such redefinition do not work.  The problem is that we
try to catch redefinition by using a special `__newindex()` metamethod
for the `LNVL.ScriptEnvironment` table.  This fails though bceause the
metamethod is only triggered when defining a key which does not
already exist within that table.  By the time we load any scripts all
of the reserved keywords exist as keys in that table due to our use of
`LNVL.CreateConstructorAlias()` and `LNVL.CreateFunctionAlias()`.

This patch addresses the problem by making the following changes:

1. The `ReservedKeywords` table now has new values.  The keys are
still strings naming the keywords, but the values are now strings
which name that type that keyword should have in the script
environment, e.g. this should always be true:

    type(LNVL.ScriptEnvironment[k]) == ReservedKeywords[k]

2. The new function `checkReservedKeywordTypes()` runs after each
dialog script we load to ensure that the condition above holds true.

3. We add a new example script demonstrating how redefining reserved
keywords causes a fatal error.

It must be noted that the method we use is not bullet-proof.  A script
could still, for example, redefine `Character` to be a different
function instead of the constructor but `checkReservedKeywordTypes()`
would not detect this as an error.