[Test] Run Luacheck on all source code
This patch first modifies the rule for running Luacheck on the core
LNVL code by adding `src/settings.lua.example` to the list of files it
should check.

Then the patch adds a new Tup rule which runs Luacheck on `main.lua`
and everything in the `examples/` directory.  This invocation of
Luacheck is kept separate because we need to run Luacheck with many
flags to loosen its standards, which we do not want to use on the core
LNVL code.  For example, for all of the example files Luacheck must be
told about read-only globals like `Character`, `Scene`, and so on, and
the same goes for the globals defined within each example,
e.g. `START`.  We ignore all warnings related to `password` because we
set that variable in `main.lua` and use it for testing shared state.