[Test] Add a Luacheck configuration for the example scripts
This patch adds a second configuration file for Luacheck which we use
only when running Luacheck on `examples/`.  The scripts in that
directory define many top-level globals (which this new configuration
ignores) and makes references to short-cut functions in LNVL which
Luacheck is unaware of (which this configuration also ignores).

With this in place we can run

    $ luacheck --config .luacheckrc-examples src/ examples/

to run Luacheck on the example scripts with appropriate settings.